project-root/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==5.4.0
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
bcrypt==4.1.3
beautifulsoup4==4.10.0
bidict==0.23.1
bleach==6.2.0
blinker==1.8.2
boto3==1.34.151
botocore==1.34.151
breadability==0.1.20
cachetools==5.5.0
certifi==2023.7.22
cffi==1.16.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cryptography==42.0.5
cycler==0.12.1
debugpy==1.8.8
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.8
distro==1.9.0
Django==5.0.2
django-filter==23.5
django-import-export==3.3.7
docopt==0.6.2
et-xmlfile==1.1.0
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.13.1
Flask==2.1.1
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.6.0
Flask-pymysql==0.2.3
Flask-ReCaptcha==0.4.2
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
fonttools==4.54.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
graphviz==0.20.3
greenlet==3.0.3
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==2.8
ipykernel==6.29.5
ipython==8.29.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.2
Jinja2==3.1.2
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
json5==0.9.27
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
lxml==5.2.2
markdown-it-py==3.0.0
MarkupPy==1.14
MarkupSafe==2.1.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
multidict==6.0.5
narwhals==1.5.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.6.3
notebook==7.2.2
notebook_shim==0.2.4
numpy==2.1.0
odfpy==1.4.1
openai==0.28.0
openpyxl==3.1.2
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandas-stubs==2.2.2.240807
pandocfilters==1.5.1
parso==0.8.4
pillow==10.4.0
platformdirs==4.2.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
protobuf==5.27.3
psutil==6.1.0
psycopg2==2.9.9
pure_eval==0.2.3
pyarrow==17.0.0
pycountry==24.6.1
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.8.0
PyMySQL==1.0.3
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-engineio==4.9.1
python-json-logger==2.0.7
python-socketio==5.11.3
pytz==2024.1
pywin32==308
pywinpty==2.0.14
PyYAML==6.0.1
pyzmq==26.2.0
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.20.0
s3transfer==0.10.2
scikit-learn==1.5.2
scipy==1.14.1
Send2Trash==1.8.3
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
sqlparse==0.4.4
stack-data==0.6.3
streamlit==1.37.1
sumy==0.11.0
sympy==1.13.1
tablib==3.5.0
tenacity==8.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.4.0
toml==0.10.2
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
types-python-dateutil==2.9.0.20241003
types-pytz==2024.1.0.20240417
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==2.1
uri-template==1.3.0
urllib3==1.25.11
virtualenv==20.25.1
watchdog==4.0.2
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==2.0.3
widgetsnbextension==4.0.13
wsproto==1.2.0
xlrd==2.0.1
xlwt==1.3.0
yarl==1.9.4
